refactor(token): convert tests to TypeScript
