+++ src/setupTests.js
@@ -1,4 +1,4 @@
-import { shouldScheduleUpdate } from './'
+import { shouldScheduleUpdate } from '.'
 
 afterAll(() => {
   shouldScheduleUpdate(false)
